chore: parking lot code

+++ src/main/java/org/example/parkinglot/strategies/feecalculation/FeeCalculationStrategy.java
@@ -3,5 +3,5 @@ package org.example.parkinglot.strategies.feecalculation;
 import org.example.parkinglot.models.Ticket;
 
 public interface FeeCalculationStrategy {
-    double calculateFee(Ticket ticket);
+    Long calculateFee(Ticket ticket);
 }
