+++ src/main/java/org/example/parkinglot/strategies/feecalculation/FlatFeeCalculationStrategy.java
@@ -4,10 +4,10 @@ import org.example.parkinglot.models.Ticket;
 
 public class FlatFeeCalculationStrategy implements FeeCalculationStrategy {
     @Override
-    public double calculateFee(Ticket ticket) {
+    public Long calculateFee(Ticket ticket) {
         // TODO:: Implement me.
         // A flat rate of ₹100 for any vehicle regardless of duration.
         // Plug it into the parking lot and test.
-        return 0;
+        return 0L;
     }
 }
